+++ Backpack-SwiftUI/Calendar/Classes/Range/RangeCalendarContainer.swift
@@ -119,45 +119,3 @@ struct RangeCalendarContainer_Previews: PreviewProvider {
         )
     }
 }
-
-struct SingleCalendarContainer<MonthHeader: View>: View {
-    @Binding var selection: Date?
-    let calendar: Calendar
-    let validRange: ClosedRange<Date>
-    @ViewBuilder let monthHeader: (_ monthDate: Date) -> MonthHeader
-    
-    private func handleSelection(_ date: Date) {
-        selection = date
-    }
-    
-    @ViewBuilder
-    private func makeDayCell(_ dayDate: Date) -> some View {
-        if selection == dayDate {
-            CalendarSelectableCell {
-                SingleSelectedCell(calendar: calendar, date: dayDate)
-            } onSelection: {
-                handleSelection(dayDate)
-            }
-        } else {
-            CalendarSelectableCell {
-                DefaultSelectionCell(calendar: calendar, date: dayDate)
-            } onSelection: {
-                handleSelection(dayDate)
-            }
-        }
-    }
-    
-    // swiftlint:disable all
-    var body: some View {
-        CalendarContainer(calendar: calendar, validRange: validRange) { month in
-            monthHeader(month)
-            CalendarMonthGrid(
-                monthDate: month,
-                calendar: calendar,
-                validRange: validRange,
-                dayCell: makeDayCell) { _, _ in
-                    DefaultEmptyCalendarDayCell()
-                }
-        }
-    }
-}
